chore: fix facility lfo uischema issues
