+++ bciers/apps/administration/app/data/jsonSchema/facilitiesSfo.ts
@@ -1,6 +1,6 @@
-import SectionFieldTemplate from "@bciers/components/form/fields/SectionFieldTemplate";
 import { RJSFSchema } from "@rjsf/utils";
 import provinceOptions from "@bciers/data/provinces.json";
+import { facilitiesLfoUiSchema } from "apps/administration/app/data/jsonSchema/facilitiesLfo";
 
 const currentYear = new Date().getFullYear();
 
@@ -101,38 +101,6 @@ export const facilitiesSchemaSfo: RJSFSchema = {
   properties: {
     section1,
     section2,
-  },
-};
-
-export const facilitiesLfoUiSchema = {
-  "ui:FieldTemplate": SectionFieldTemplate,
-  "ui: options": {
-    label: false,
-  },
-  section1: {
-    "ui:FieldTemplate": SectionFieldTemplate,
-    type: {
-      "ui:widget": "ComboBox",
-    },
-    is_current_year: {
-      "ui:widget": "ToggleWidget",
-    },
-    starting_date: {
-      "ui:widget": "DateWidget",
-      "ui:options": {
-        maxDate: new Date(),
-        minDate: new Date(currentYear - 1, 0, 1),
-      },
-    },
-  },
-  section2: {
-    "ui:FieldTemplate": SectionFieldTemplate,
-    province: {
-      "ui:widget": "ComboBox",
-    },
-    postal_code: {
-      "ui:widget": "PostalCodeWidget",
-    },
   },
 };
